// src/shared/types/db.ts
export namespace DB {
	export type AimeUser = {
		id: number
		username: string
		email: string
		password: string
		permissions: number
		created_date: string
		last_login_date: string
		suspend_expire_time: string
		access_code?: string
	}
// ...
	export type Machine = {
		id: number
		arcade: number
		serial: string
		board: string | null
		game: string | null
		country: string | null
		timezone: string | null
		ota_enable: boolean | null
		memo: string | null
		is_cab: boolean | null
		data: string | null
		nickname?: string
		name?: string
		user?: number
	}
// ...
}